# api/server_swg/src/main/java/io/swagger/api/TicketsApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.46).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.NewTicket;
import io.swagger.model.Ticket;
import io.swagger.model.UpdateTicket;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2025-10-29T17:50:22.611548500-04:00[America/Toronto]")
@Validated
public interface TicketsApi {

    @Operation(summary = "Liste tous les tickets", description = "", tags={ "tickets" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "Liste des tickets", content = @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Ticket.class)))) })
    @RequestMapping(value = "/tickets",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<Ticket>> ticketsGet();


    @Operation(summary = "Récupère un ticket par ID", description = "", tags={ "tickets" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "Ticket trouvé", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Ticket.class))),
        
        @ApiResponse(responseCode = "404", description = "Ticket introuvable") })
    @RequestMapping(value = "/tickets/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Ticket> ticketsIdGet(@Parameter(in = ParameterIn.PATH, description = "", required=true, schema=@Schema()) @PathVariable("id") Integer id);


    @Operation(summary = "Met à jour un ticket (statut, priorité, assignation)", description = "", tags={ "tickets" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "Ticket mis à jour", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Ticket.class))) })
    @RequestMapping(value = "/tickets/{id}",
        produces = { "application/json" }, 
        consumes = { "application/json" }, 
        method = RequestMethod.PUT)
    ResponseEntity<Ticket> ticketsIdPut(@Parameter(in = ParameterIn.PATH, description = "", required=true, schema=@Schema()) @PathVariable("id") Integer id, @Parameter(in = ParameterIn.DEFAULT, description = "", required=true, schema=@Schema()) @Valid @RequestBody UpdateTicket body);


    @Operation(summary = "Crée un nouveau ticket", description = "", tags={ "tickets" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "201", description = "Ticket créé", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Ticket.class))) })
    @RequestMapping(value = "/tickets",
        produces = { "application/json" }, 
        consumes = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<Ticket> ticketsPost(@Parameter(in = ParameterIn.DEFAULT, description = "", required=true, schema=@Schema()) @Valid @RequestBody NewTicket body);

}