# api/server_swg/src/main/java/org/openapitools/api/TicketsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.8.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.NewTicket;
import org.openapitools.model.Ticket;
import org.openapitools.model.UpdateTicket;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-10-29T17:51:12.972379500-04:00[America/Toronto]", comments = "Generator version: 7.8.0")
@Validated
@Tag(name = "tickets", description = "Opérations sur les tickets")
public interface TicketsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /tickets : Liste tous les tickets
     *
     * @return Liste des tickets (status code 200)
     */
    @Operation(
        operationId = "ticketsGet",
        summary = "Liste tous les tickets",
        tags = { "tickets" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Liste des tickets", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Ticket.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/tickets",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Ticket>> ticketsGet(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"title\" : \"title\", \"priority\" : \"BASSE\", \"assignedUser\" : { \"role\" : \"role\", \"name\" : \"name\", \"userID\" : 6, \"email\" : \"email\" }, \"descriptions\" : [ { \"addedBy\" : { \"role\" : \"role\", \"name\" : \"name\", \"userID\" : 6, \"email\" : \"email\" }, \"id\" : 1, \"text\" : \"text\", \"type\" : \"TEXTE\", \"uri\" : \"uri\" }, { \"addedBy\" : { \"role\" : \"role\", \"name\" : \"name\", \"userID\" : 6, \"email\" : \"email\" }, \"id\" : 1, \"text\" : \"text\", \"type\" : \"TEXTE\", \"uri\" : \"uri\" } ], \"ticketID\" : 0, \"status\" : \"OUVERT\" }, { \"title\" : \"title\", \"priority\" : \"BASSE\", \"assignedUser\" : { \"role\" : \"role\", \"name\" : \"name\", \"userID\" : 6, \"email\" : \"email\" }, \"descriptions\" : [ { \"addedBy\" : { \"role\" : \"role\", \"name\" : \"name\", \"userID\" : 6, \"email\" : \"email\" }, \"id\" : 1, \"text\" : \"text\", \"type\" : \"TEXTE\", \"uri\" : \"uri\" }, { \"addedBy\" : { \"role\" : \"role\", \"name\" : \"name\", \"userID\" : 6, \"email\" : \"email\" }, \"id\" : 1, \"text\" : \"text\", \"type\" : \"TEXTE\", \"uri\" : \"uri\" } ], \"ticketID\" : 0, \"status\" : \"OUVERT\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /tickets/{id} : Récupère un ticket par ID
     *
     * @param id  (required)
     * @return Ticket trouvé (status code 200)
     *         or Ticket introuvable (status code 404)
     */
    @Operation(
        operationId = "ticketsIdGet",
        summary = "Récupère un ticket par ID",
        tags = { "tickets" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Ticket trouvé", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Ticket.class))
            }),
            @ApiResponse(responseCode = "404", description = "Ticket introuvable")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/tickets/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Ticket> ticketsIdGet(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"title\" : \"title\", \"priority\" : \"BASSE\", \"assignedUser\" : { \"role\" : \"role\", \"name\" : \"name\", \"userID\" : 6, \"email\" : \"email\" }, \"descriptions\" : [ { \"addedBy\" : { \"role\" : \"role\", \"name\" : \"name\", \"userID\" : 6, \"email\" : \"email\" }, \"id\" : 1, \"text\" : \"text\", \"type\" : \"TEXTE\", \"uri\" : \"uri\" }, { \"addedBy\" : { \"role\" : \"role\", \"name\" : \"name\", \"userID\" : 6, \"email\" : \"email\" }, \"id\" : 1, \"text\" : \"text\", \"type\" : \"TEXTE\", \"uri\" : \"uri\" } ], \"ticketID\" : 0, \"status\" : \"OUVERT\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /tickets/{id} : Met à jour un ticket (statut, priorité, assignation)
     *
     * @param id  (required)
     * @param updateTicket  (required)
     * @return Ticket mis à jour (status code 200)
     */
    @Operation(
        operationId = "ticketsIdPut",
        summary = "Met à jour un ticket (statut, priorité, assignation)",
        tags = { "tickets" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Ticket mis à jour", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Ticket.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/tickets/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Ticket> ticketsIdPut(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "UpdateTicket", description = "", required = true) @Valid @RequestBody UpdateTicket updateTicket
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"title\" : \"title\", \"priority\" : \"BASSE\", \"assignedUser\" : { \"role\" : \"role\", \"name\" : \"name\", \"userID\" : 6, \"email\" : \"email\" }, \"descriptions\" : [ { \"addedBy\" : { \"role\" : \"role\", \"name\" : \"name\", \"userID\" : 6, \"email\" : \"email\" }, \"id\" : 1, \"text\" : \"text\", \"type\" : \"TEXTE\", \"uri\" : \"uri\" }, { \"addedBy\" : { \"role\" : \"role\", \"name\" : \"name\", \"userID\" : 6, \"email\" : \"email\" }, \"id\" : 1, \"text\" : \"text\", \"type\" : \"TEXTE\", \"uri\" : \"uri\" } ], \"ticketID\" : 0, \"status\" : \"OUVERT\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /tickets : Crée un nouveau ticket
     *
     * @param newTicket  (required)
     * @return Ticket créé (status code 201)
     */
    @Operation(
        operationId = "ticketsPost",
        summary = "Crée un nouveau ticket",
        tags = { "tickets" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Ticket créé", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Ticket.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/tickets",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Ticket> ticketsPost(
        @Parameter(name = "NewTicket", description = "", required = true) @Valid @RequestBody NewTicket newTicket
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"title\" : \"title\", \"priority\" : \"BASSE\", \"assignedUser\" : { \"role\" : \"role\", \"name\" : \"name\", \"userID\" : 6, \"email\" : \"email\" }, \"descriptions\" : [ { \"addedBy\" : { \"role\" : \"role\", \"name\" : \"name\", \"userID\" : 6, \"email\" : \"email\" }, \"id\" : 1, \"text\" : \"text\", \"type\" : \"TEXTE\", \"uri\" : \"uri\" }, { \"addedBy\" : { \"role\" : \"role\", \"name\" : \"name\", \"userID\" : 6, \"email\" : \"email\" }, \"id\" : 1, \"text\" : \"text\", \"type\" : \"TEXTE\", \"uri\" : \"uri\" } ], \"ticketID\" : 0, \"status\" : \"OUVERT\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}